Moscrypt/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
cryptography==40.0.2
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==2.2.3
psycopg2-binary==2.9.5
Flask-WTF==1.1.1
alembic==1.10.3 